pyproject.toml
[project]
name = "apart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
